requirements.txt
curl-cffi==0.7.4
beautifulsoup4==4.12.3
lxml==5.2.2
aiosqlite==0.20.0
tenacity==8.3.0
rich==13.7.1
python-dotenv==1.0.1
httpx==0.27.0
